Chore: "Refined functionality with fragment approach for the app"
