+++ app/src/main/java/com/cotiledon/mobilApp/ui/activities/MainAppApproach/MainAppHomeActivity.kt
@@ -1,7 +1,6 @@
-package com.cotiledon.mobilApp.ui.activities
+package com.cotiledon.mobilApp.ui.activities.MainAppApproach
 
 import android.os.Bundle
-import android.view.Menu
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
