+++ app/src/main/java/com/cotiledon/mobilApp/ui/activities/UserRegistrationActivity.kt
@@ -19,6 +19,7 @@ import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.lifecycleScope
 import com.cotiledon.mobilApp.R
+import com.cotiledon.mobilApp.ui.activities._old.HomeActivity
 import com.cotiledon.mobilApp.ui.dataClasses.UserRegistration
 import com.cotiledon.mobilApp.ui.retrofit.RetrofitUserClient
 import kotlinx.coroutines.Dispatchers
