+++ app/src/main/java/com/cotiledon/mobilApp/ui/fragments/HomeFragment.kt
@@ -20,7 +20,7 @@ class HomeFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the fragment's layout
-        return inflater.inflate(R.layout.activity_home_fragment, container, false)
+        return inflater.inflate(R.layout.fragment_home, container, false)
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
