+++ app/src/main/res/layout/activity_catalog.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.activities.CatalogActivity">
+    tools:context=".ui.activities._old.CatalogActivity">
 
     <FrameLayout
         android:id="@+id/CatalogTopFrmBanner"
