+++ app/src/main/res/layout/activity_main_app.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.activities.MainAppHomeActivity">
+    tools:context=".ui.activities.MainAppApproach.MainAppHomeActivity">
 
     <!-- ScrollView for the main content -->
     <ScrollView
