+++ app/src/main/res/layout/activity_main_app_profile.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.activities.MainAppCategoriesActivity">
+    tools:context=".ui.activities.MainAppApproach.MainAppCategoriesActivity">
 
     <!-- Comienzo de Header -->
     <View
