+++ app/src/main/res/layout/activity_main_container.xml
@@ -17,7 +17,7 @@
     <com.google.android.material.bottomnavigation.BottomNavigationView
         android:id="@+id/bottom_navigation"
         android:layout_width="match_parent"
-        android:layout_height="79dp"
+        android:layout_height="80dp"
         android:layout_gravity="bottom"
         android:background="@drawable/menu_background"
         android:elevation="8dp"
