+++ app/src/main/res/layout/activity_payment_details.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.activities.PaymentDetailsActivity">
+    tools:context=".ui.activities._old.PaymentDetailsActivity">
 
     <androidx.constraintlayout.widget.Guideline
         android:id="@+id/gdHomeTop"
