+++ app/src/main/res/layout/activity_shopping_cart.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#E4F5D9"
-    tools:context=".ui.activities.SignInActivity_unatended">
+    tools:context=".ui.activities._old.SignInActivity_unatended">
 
     <androidx.constraintlayout.widget.Guideline
         android:id="@+id/gdlnBottom"
