+++ app/src/main/res/layout/fragment_categories.xml
@@ -31,6 +31,7 @@
         app:layout_constraintTop_toBottomOf="@+id/app_upper_banner">
 
         <ImageView
+            android:id="@+id/search_icon"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_gravity="center_vertical"
@@ -38,6 +39,7 @@
             android:src="@drawable/search" />
 
         <EditText
+            android:id="@+id/search_edit_text"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
             android:layout_marginStart="20dp"
@@ -50,6 +52,7 @@
             android:textSize="18sp" />
 
         <ImageView
+            android:id="@+id/camera_button"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_gravity="center_vertical"
