+++ app/src/main/res/layout/fragment_home.xml
@@ -180,6 +180,8 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_marginTop="16dp"
+            android:layout_marginBottom="16dp"
+            app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/activity_main_banner_down">
 
             <!-- Guidelines for alignment -->
@@ -202,8 +204,8 @@
                 android:id="@+id/truck_icon"
                 android:layout_width="50dp"
                 android:layout_height="50dp"
-                android:src="@drawable/truck"
                 android:layout_marginEnd="12dp"
+                android:src="@drawable/truck"
                 app:layout_constraintEnd_toStartOf="@id/start_guideline"
                 app:layout_constraintTop_toTopOf="parent" />
 
@@ -302,8 +304,8 @@
                 android:textColor="@color/black"
                 android:textSize="14sp"
                 android:textStyle="bold"
-                app:layout_constraintStart_toStartOf="@id/start_guideline"
                 app:layout_constraintEnd_toEndOf="@id/end_guideline"
+                app:layout_constraintStart_toStartOf="@id/start_guideline"
                 app:layout_constraintTop_toTopOf="@id/grass_icon" />
 
             <Button
